+++ main_gui.py
@@ -1,0 +1,17 @@
+from functions import *
+from Bio import SeqIO
+import pandas as pd
+
+from tkinter import *
+from tkinter import ttk
+
+window = Tk()
+
+window.title("Cuterle")
+window.config(padx=0, pady=0, bg="white")
+
+
+
+
+
+window.mainloop()
